Fix: Floating point error for big number division
Refactor: Rework of `Bytes` to incorporate binary & decimal-based values

Test: Adjusted tests accordingly

+++ src/main/nextflow/co2footprint/Parsers/TraceFileParser.groovy
@@ -43,7 +43,7 @@ class TraceFileParser {
                     value = switch (TraceRecord.FIELDS.get(key)) {
                         case 'mem' -> if (value.endsWith('B')) {
                             List<String> split = value.split(' ')
-                            Metric<BigDecimal> bytes = Bytes.of(split[0].toDouble(), split[1].dropRight(1)).scale('')
+                            Metric<BigDecimal> bytes = Bytes.of(split[0] as BigDecimal, split[1].dropRight(1)).scale('')
                             bytes.value.toLong()
                         } else {
                             value.toLong()
